Cambio de tipo Int a Byte

+++ Persistencia/Domain/Cupo.cs
@@ -8,12 +8,12 @@ namespace Persistencia.Domain
 {
     public class Cupo
     {
-        public int CupoId { get; set; }
+        public byte CupoId { get; set; }
         public TipoCupo Tipo { get; set; }
         public bool Disponible { get; set; }
 
 
-        public int HabitacionId { get; set; }
+        public byte HabitacionId { get; set; }
         public Habitacion Habitacion { get; set; }
 
     }
