+++ Persistencia/Domain/Habitacion.cs
@@ -8,9 +8,9 @@ namespace Persistencia.Domain
 {
     public class Habitacion
     {
-        public int HabitacionId { get; set; }//Numero de la Habitacion
+        public byte HabitacionId { get; set; }//Numero de la Habitacion
         //public int Capacidad { get; set; }
-        public int Planta { get; set; }
+        public byte Planta { get; set; }
         public bool Exclusiva { get; set; }
 
         public List<Cupo> Cupos { get; set; }
