+++ Persistencia/Domain/LineaServicio.cs
@@ -9,12 +9,12 @@ namespace Persistencia.Domain
     public class LineaServicio
     {
         public int LineaServicioId { get; set; }
-        public int Cantidad { get; set; }
+        public byte Cantidad { get; set; }
         public double CostoServicio { get; set; }
         public DateTime FechaServicio { get; set; }
 
 
-        public int ServicioId { get; set; }
+        public byte ServicioId { get; set; }
         public Servicio Servicio { get; set; }
 
         public int AlojamientoId { get; set; }
